fixed wrong sprite on player, added menu and game bg & fixed warnings

+++ MapManager.cpp
@@ -1238,7 +1238,7 @@ void CheckpointTile::Init() {
 
 void CheckpointTile::Update()
 {
-	float dt = AEFrameRateControllerGetFrameTime();
+	float dt = static_cast<float>(AEFrameRateControllerGetFrameTime());
     Tile::Update();
 	checkpointParticles.Update(dt);
 }
